refactor(e2e): separate dem-dev test target
