+++ e2e/environments/dem-dev/runtime.env
@@ -1,4 +1,4 @@
-# Temporary source-level runtime lock for the first dem-dev consumer.
+# Temporary source-level runtime lock for the closed-loop dem-dev suite.
 #
 # The intended end state is one DEM_RUNTIME_VERSION whose release contains the
 # compatible API, agent, scenario-runner, and infrastructure image digests.
