+++ e2e/tsconfig.json
@@ -6,5 +6,5 @@
     "resolveJsonModule": true,
     "types": ["node", "@playwright/test"]
   },
-  "include": ["./**/*.ts", "../playwright.config.ts"]
+  "include": ["./**/*.ts", "../playwright.dem-dev.config.ts"]
 }
