+++ package.json
@@ -13,9 +13,9 @@
     "danger:pr": "yarn danger pr --dangerfile danger/dangerfile.ts",
     "dev:msw": "REACT_APP_MSW=true webpack -w -c webpack.config.ts --env development",
     "dev": "webpack -w -c webpack.config.ts --env development",
-    "e2e": "playwright test",
-    "e2e:update": "playwright test --update-snapshots",
-    "e2e:ui": "playwright test --ui",
+    "e2e": "playwright test --config playwright.dem-dev.config.ts",
+    "e2e:update": "playwright test --config playwright.dem-dev.config.ts --update-snapshots",
+    "e2e:ui": "playwright test --config playwright.dem-dev.config.ts --ui",
     "e2e:install": "playwright install chromium",
     "e2e:dem": "bash ./scripts/dem-dev-e2e.sh run",
     "e2e:dem:configure": "bash ./scripts/dem-dev-e2e.sh configure",
